memory-product-staging/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
psycopg2-binary==2.9.11
redis==7.3.0
requests==2.31.0
pydantic==2.12.5
psutil==6.1.1
posthog==3.5.0
